+++ doc/version_history.rst
@@ -10,6 +10,7 @@ v1.16.0
 =======
 * Remove backward compatibility with ts_xml 20.2.
 * Add do_setPowerManagementMode command.
+* Fix conda recipe.
 
 Requires: